% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioaccumulation_factor.R
\name{bioaccumulation_factor}
\alias{bioaccumulation_factor}
\title{Bioaccumulation Factor}
\usage{
bioaccumulation_factor(plant_metal, soil_metal)
}
\arguments{
\item{plant_metal}{Metal concentration in plant}

\item{soil_metal}{Metal concentration in soil}
}
\value{
numeric
}
\description{
Calculates metal accumulation
from soil into plant tissue.
}
