% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keystone_taxa.R
\name{keystone_taxa}
\alias{keystone_taxa}
\title{Detect Keystone Taxa}
\usage{
keystone_taxa(network)
}
\arguments{
\item{network}{igraph network}
}
\value{
dataframe
}
\description{
Detect Keystone Taxa
}
