% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metal_correlate.R
\name{metal_correlate}
\alias{metal_correlate}
\title{Correlation Between Metals and Taxa}
\usage{
metal_correlate(otu, metal, method = "spearman")
}
\arguments{
\item{otu}{OTU abundance matrix}

\item{metal}{Metal concentration data}

\item{method}{Correlation method}
}
\value{
Data frame
}
\description{
Correlation Between Metals and Taxa
}
