% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metal_heatmap.R
\name{metal_heatmap}
\alias{metal_heatmap}
\title{Metal Correlation Heatmap}
\usage{
metal_heatmap(correlation_table)
}
\arguments{
\item{correlation_table}{Output from metal_correlate()}
}
\description{
Metal Correlation Heatmap
}
