% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metal_indicator_taxa.R
\name{metal_indicator_taxa}
\alias{metal_indicator_taxa}
\title{Indicator Taxa Analysis}
\usage{
metal_indicator_taxa(otu, group)
}
\arguments{
\item{otu}{OTU abundance matrix (samples × taxa)}

\item{group}{treatment vector}
}
\value{
indicspecies object
}
\description{
Identify taxa associated with heavy metal treatments.
}
