% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metal_network.R
\name{metal_network}
\alias{metal_network}
\title{Metal-Microbe Network}
\usage{
metal_network(correlations, cutoff = 0.6)
}
\arguments{
\item{correlations}{Output from metal_correlate()}

\item{cutoff}{Correlation cutoff}
}
\value{
igraph object
}
\description{
Build network from significant
metal-taxa associations.
}
