% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metal_tolerance_index.R
\name{metal_tolerance_index}
\alias{metal_tolerance_index}
\title{Metal Tolerance Index}
\usage{
metal_tolerance_index(stress, control)
}
\arguments{
\item{stress}{abundance under contamination}

\item{control}{abundance under control}
}
\value{
numeric
}
\description{
Calculates abundance ratio between
contaminated and control conditions.
}
