% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remediation_index.R
\name{remediation_index}
\alias{remediation_index}
\title{Remediation Efficiency Index}
\usage{
remediation_index(control, treated)
}
\arguments{
\item{control}{Initial metal concentration}

\item{treated}{Final metal concentration}
}
\value{
numeric
}
\description{
Calculates percentage removal of a metal
after remediation treatment.
}
