% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translocation_factor.R
\name{translocation_factor}
\alias{translocation_factor}
\title{Translocation Factor}
\usage{
translocation_factor(shoot_metal, root_metal)
}
\arguments{
\item{shoot_metal}{Metal concentration in shoot}

\item{root_metal}{Metal concentration in root}
}
\value{
numeric
}
\description{
Measures movement of metals
from roots to shoots.
}
